test : 메서드 위치 이동
